Add a function setting to parse the display value
The display value being "name" was not ideal for me. I wanted to
display the full name of the user as well as their user name in the
dropdown, but I only wanted to output '@' + id into the text area the
same as it is on Twitter. Our server already parses mentions in this
format so it makes sense for us to be able to override this.